test: extract loopback accept-queue-saturation helper into ic-test-utilities-net
PR #10803 introduced a trick to test connect-timeout behavior without
network egress: bind a loopback listener with a minimal backlog
(`listen(1)`), saturate its accept queue so the kernel silently drops
further SYNs, and let a fresh connect hang until the timeout fires. The
same block was then copied into the bitcoin adapter to drop its
`requires-network` tag, leaving two near-identical inline copies.

Extract it into a new tokio-only, testonly crate `ic-test-utilities-net`
exposing `saturated_loopback_listener() -> SaturatedTcpListener`. The
guard owns both the listener and the blocker connections, so the address
stays unconnectable for as long as the guard is alive.

Both call sites now just do:

    let _saturated = ic_test_utilities_net::saturated_loopback_listener().await;
    let addr = _saturated.addr();

- rs/https_outcalls/adapter/tests/server_test.rs (`test_connect_timeout`)
- rs/bitcoin/adapter/src/stream.rs (`initialization_times_out_after_five_seconds`)
