+++ rs/bitcoin/adapter/Cargo.toml
@@ -47,6 +47,7 @@ ic-btc-replica-types = { path = "../replica_types" }
 ic-crypto-test-utils-reproducible-rng = { path = "../../crypto/test_utils/reproducible_rng" }
 ic-interfaces-adapter-client = { path = "../../interfaces/adapter_client" }
 ic-test-utilities-logger = { path = "../../test_utilities/logger" }
+ic-test-utilities-net = { path = "../../test_utilities/net" }
 sha2 = { workspace = true }
 scrypt = { version = "0.11.0", default-features = false }
 tempfile = { workspace = true }
