+++ rs/https_outcalls/adapter/BUILD.bazel
@@ -241,6 +241,7 @@ rust_test_suite(
         "//rs/https_outcalls/service",
         "//rs/monitoring/logger",
         "//rs/monitoring/metrics",
+        "//rs/test_utilities/net",
         "@crate_index//:async-stream",
         "@crate_index//:bytes",
         "@crate_index//:futures",
@@ -284,6 +285,7 @@ rust_test_suite(
         "//rs/https_outcalls/service",
         "//rs/monitoring/logger",
         "//rs/monitoring/metrics",
+        "//rs/test_utilities/net",
         "@crate_index//:async-stream",
         "@crate_index//:bytes",
         "@crate_index//:futures",
