+++ rs/test_utilities/net/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@rules_rust//rust:defs.bzl", "rust_library")
+
+package(default_visibility = ["//visibility:public"])
+
+rust_library(
+    name = "net",
+    testonly = True,
+    srcs = glob(["src/**/*.rs"]),
+    crate_name = "ic_test_utilities_net",
+    version = "0.9.0",
+    deps = [
+        "@crate_index//:tokio",
+    ],
+)
